Add DUMP request support for the server, this can be used to get a total snapshot of the tree, complete with values and everything - the functionality is not used by SeqZap yet so it might change.

+++ trunk/server/c/applications/machine_automation/process.c
@@ -6,7 +6,7 @@
 #include "common.h"
 
 static char EMPTY[] = "";
-static int g_numberOfProcesses = 0;
+static size_t g_numberOfProcesses = 0;
 static struct Process * g_processes = NULL;
 
 void process_resize_internal_buffer ( size_t _newSize )
